pyproject.toml
[project]
name = "contri-backend"
version = "0.1.0"
description = "Backend API for Contri Mobile"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "celery>=5.3.6",
    "redis>=5.0.3",
    "httpx-oauth>=0.14.1",
    "python-socketio[asyncio_client]>=5.11.2",
    "sqladmin>=0.16.1",
    "email-validator>=2.1.1",
    "sentry-sdk>=1.45.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

